feat: remove file-based input (#184)
This PR removes necessity to have a `.addonmatrix` file in the root of
the directory to configure which versions of Splunk need to be tested.

I do not believe that file was ever changed so there is no need as of
now to provide it as an input parameter in `action.yml` but this option
is on the table in case it will be needed in the future.

I believe it's safe to mark it as a breaking change in case someone uses
non-default value in `.addonmatrix` file. Please reach out and we will
expose `--splunkfeatures` parameter through `action.yml` file.

BREAKING CHANGE: remove file-based input